feat(#60): 스터디 상세 조회 API 구현
연관되어 있는 API 구현 필요(공지, 투표, 멤버, 피드)

+++ app/src/main/java/com/ssafy/neegongnaegong/data/datasource/network/NetworkStudiesDataSource.kt
@@ -12,7 +12,7 @@ interface NetworkStudiesDataSource {
 
     suspend fun createStudies(request: CreateStudiesRequest): Flow<Unit>
 
-    suspend fun getStudies(studyGroupId: Int): Flow<StudiesResponse>
+    suspend fun getStudies(studyGroupId: Long): Flow<StudiesResponse>
 
     suspend fun updateStudies(
         studyGroupId: Long,
